refactor: consolidate into a single moderator class
