+++ packages/aila/package.json
@@ -43,6 +43,7 @@
     "immer": "^10.1.1",
     "jsonrepair": "^3.8.0",
     "openai": "^4.96.0",
+    "openapi-fetch": "^0.15.0",
     "p-limit": "^7.2.0",
     "remeda": "^1.29.0",
     "superjson": "^1.9.1",
