+++ packages/core/package.json
@@ -8,8 +8,6 @@
   "main": "./index.ts",
   "scripts": {
     "admin": "open http://localhost:8288",
-    "codegen:moderation": "curl -s \"https://moderation-api.thenational.academy/openapi.json?x-vercel-protection-bypass=${VERCEL_AUTOMATION_BYPASS_SECRET}\" | openapi-typescript /dev/stdin -o src/generated/moderation-api.d.ts",
-    "codegen": "pnpm codegen:moderation",
     "check": "tsc --noEmit",
     "clean": "rm -rf .turbo node_modules",
     "debug": "open https://localhost:2525/api/inngest",
@@ -63,7 +61,6 @@
     "murmurhash": "^2.0.1",
     "obscenity": "^0.1.4",
     "openai": "^4.96.0",
-    "openapi-fetch": "^0.15.0",
     "tiny-invariant": "^1.3.1",
     "ts-dedent": "^2.2.0",
     "ts-md5": "^1.3.1",
