--- packages/core/src/services/moderation/index.ts
@@ -1,7 +1,0 @@
-export {
-  OakModerationServiceClient,
-  OakModerationServiceError,
-  type OakModerationServiceClientConfig,
-  type ModerationServiceResponse,
-  type ModerationCategory,
-} from "./OakModerationServiceClient";
